reorganize seeds and reseed
