--- data/dev/foods.js
@@ -1,9 +1,0 @@
-module.exports = [
-  {name: 'Banana', calories: 150},
-  {name: 'Bagel Bites - Four Cheese', calories: 650},
-  {name: 'Chicken Burrito', calories: 800},
-  {name: 'Yogurt', calories: 550},
-  {name: 'Gum', calories: 50},
-  {name: 'Cheese', calories: 400},
-  {name: 'Apple', calories: 220}
-]
