--- data/dev/mealFoods.js
@@ -1,14 +1,0 @@
-module.exports = [
-  {meal_id: 1, food_id: 1},
-  {meal_id: 1, food_id: 4},
-  {meal_id: 1, food_id: 7},
-  {meal_id: 2, food_id: 1},
-  {meal_id: 2, food_id: 5},
-  {meal_id: 2, food_id: 6},
-  {meal_id: 3, food_id: 2},
-  {meal_id: 3, food_id: 3},
-  {meal_id: 3, food_id: 7},
-  {meal_id: 4, food_id: 1},
-  {meal_id: 4, food_id: 2},
-  {meal_id: 4, food_id: 3}
-]
