--- data/dev/meals.js
@@ -1,6 +1,0 @@
-module.exports = [
-  {name: 'Breakfast'},
-  {name: 'Snack'},
-  {name: 'Lunch'},
-  {name: 'Dinner'}
-]
