+++ data/meals.js
@@ -1,6 +1,6 @@
 module.exports = [
-  {id: 1, name: 'Breakfast'},
-  {id: 2, name: 'Snack'},
-  {id: 3, name: 'Lunch'},
-  {id: 4, name: 'Dinner'}
+  {name: 'Breakfast'},
+  {name: 'Snack'},
+  {name: 'Lunch'},
+  {name: 'Dinner'}
 ]
