+++ data/test/meals.js
@@ -1,0 +1,6 @@
+module.exports = [
+  {id: 1, name: 'Breakfast'},
+  {id: 2, name: 'Snack'},
+  {id: 3, name: 'Lunch'},
+  {id: 4, name: 'Dinner'}
+]
